Began implementing async non-blocking application-server interface
The application-server implements communication between the mygcc-api
server running on the localhost and the application. This is challenging
because not only are the application and server on different threads,
they are on different processes completely. As such the applcation must
make sure that the server is running correctly before sending requests.
Turns out designing the application-server interface is harder than it
sounds.